[LIbrary] Add motor interrupt pin to button Add motion constants according to external controller testing result Add complete flag to controller communication packge sending Change controller communication slaveID to broadcast ID 0xFFFF Adjust silence time to negative values and prioritize shorter silence time
[Task]
Add one-shot timer as task ID 0 and its and its interaction function
Disable all error checking except value_not_set due to hardware shortage

[System]
Add motion parameter calculation needed to pass to controller
Update state transition condition using motro interrupt pin level
Update state transition actions
Add main source file

[MXCube]
Change ADC2 all channels to Injected conversion mode

+++ Core/Src/sensor.c
@@ -48,7 +48,7 @@ void Sensor_UpdateValue(uint32_t* pui32Raw)
 SensorData_t* Sensor_GetData(void)
 {
     /* Convert data */
-    data.pressure = (22 * data_raw[0] - 68250) / 2457.0f * KPA_TO_CMH2O_FACTOR;
+    data.pressure = (float)(22 * data_raw[0] - 68250) / 2457.0f * KPA_TO_CMH2O_FACTOR;
     data.angle = data_raw[1];
     data.ris = data_raw[2];
     data.lis = data_raw[3];
